def ryerson_letter_grade(ptc):
Given the percentage grade for the course, calculate and return the letter grade that would appear in the Ryerson grades transcript, as defined on the page Ryerson Grade Scales. The letter grade should be returned as a string that consists of the uppercase letter followed by the possible modifier '+' or '-'. The function should work correctly for all values of pct from 0 to 150.