def all_cyclic_shifts(text):
Given a text string, create and return the list of all its possible cyclic shifts, that is, strings of the same length that can be constructed from text by moving some prefix of that string to the end of that string. For example, the cyclic shifts of 'hello' would be 'hello', 'elloh', 'llohe', 'lohel', and 'ohell'. Once again, zero is always a possibility in programming, so every string is trivially its own cyclic shift by moving the empty prefix to the end. (Always keep in mind the important principle that unlike most systems and processes designed for humans in the real world, in computer programming zero is a legitimate possibility that your code must be prepared for, usually by doing nothing the same way as in here, since doing nothing is the correct thing to do in that situation. But as we know, sometimes it can be so very hard in life to do nothing...)