def squares_intersect(s1,s2):
A square on the two-dimensional plane can be defined as a tuple (x, y, s) where (x, y) are the coordinates of its bottom left corner and s is the length of the side of the square. Given two squares (x1, y1, s1) and (x2, y2, s2), this function should determine whether these two squares intersect, that is, their areas have at least one point in common, even if that one point is merely the shared corner point when these two squares are placed corner touching corner. This function should not contain any loops or comprehensions of any kind, but should compute the result using only conditional statements.