def detab(text, n = 8, sub = ' '):
In the detabbing process of converting tab characters '\t' to ordinary whitespaces, each tab character is replaced by a suitable number of whitespace characters so that the next character is placed at a position that is exactly divisible by n. However, if the next character is already in a position that is divisible by n, another n whitespace characters are appended to the result.