def reverse_reversed(items):
Create and return a new list that contains the items in reverse, but so that whenever each item is itself a list, its elements are also reversed. This reversal of sublists must keep going on all the way down, no matter how deep the nesting of these lists, so you necessarily have to use recursion to solve this problem. The base case handles any argument that is not a list. When items is a list (use the Python function type or the isinstance operator to check this), recursively reverse the elements that this nested list contains. (List comprehensions might come handy in solving this problem.) Note that this function must create and return a new list to represent the result, and should not rearrange or otherwise touch the contents of the original list.