def hand_is_badugi(hand):
In the exotic draw poker variation of badugi, the four cards held by the player form the titular and much sought after badugi if no two cards in that hand have the same suit or the same rank. Given the hand as a list of four elements so that each card is encoded as a tuple of strings (rank, suit) the exact same way as in the cardproblems.py example program, determine whether that hand is a badugi.