def reverse_ascending_sublists(items):
Create and return a new list that contains the same elements as the argument list items, but reversing the order of the elements inside every maximal strictly ascending sublist. This function should not modify the contents of the original list, but create and return a new list object that contains the result.