def bridge_hand_shape(hand):
In the card game of bridge, each player receives a hand that contains thirteen cards. The shape of the hand is the distribution of these cards into the four suits in the exact order of spades, hearts, diamonds, and clubs. Given a bridge hand encoded as in the example script cardproblems.py, return the list of these four numbers. For example, given a hand that contains five spades, no hearts, five diamonds and three clubs, this function should return [5, 0, 5, 3]. Note that the cards in hand can be given to your function in any order, since in this question the player has not yet manually sorted his hand. Your answer still has to list the suits in the required order.