def pancake_scramble(text):
Analogous to flipping a stack of pancakes by sticking a spatula inside the stack and flipping pancakes that rest on top of that spatula, a pancake flip of order k done for the given text string reverses the prefix of first k characters and keeps the rest of the string as it were. For example, the pancake flip of order 2 performed on the string 'ilkka' would produce the string 'likka'. The pancake flip of order 3 performed on the same string would produce 'klika'.