def reverse_vowels(text):
Given a text string, create and return a new string constructed by finding all its vowels (for simplicity, in this problem vowels are the letters in the string 'aeiouAEIOU') and reversing their order, while keeping all non-vowel characters exactly as they were in their original positions. However, to make the result look prettier, the capitalization of each vowel must be the same as that of the vowel originally in that same position. For example, reversing the vowels of 'Ilkka' should produce 'Alkki' instead of 'alkkI'.