def is_cyclops(n):
A nonnegative integer is said to be a cyclops number if it consists of an odd number of digits so that the middle (or more poetically, the "eye") digit is a zero, and all other digits of that number are nonzero. This function should determine whether its parameter integer n is a cyclops number, and accordingly return either True or False.