def scrabble_value(word, multipliers = None):
Compute the point value of the given word in Scrabble, using the standard English version letter points. The parameter multipliers is a list of integers of the same length as word, telling how much the value of the letter in each position gets multiplied by. It is guaranteed that word consists of lowercase English letters only. If multipliers == None, your function should use the default multiplier value of one for each position.