def tribonacci(n, start = (1, 1, 1)):
The famous and important Fibonacci series is defined to start with values (1, 1) and from there, each element in the sequence equals the sum of the previous two elements. However, a cute variation aptly called the Tribonacci series works otherwise the same way, but starts with a triple of values (1, 1, 1) and from there onwards, each element equals the sum of the previous three elements. Tribonacci numbers grow quite a bit faster than Fibonacci numbers, the first twenty being 1, 1, 1, 3, 5, 9, 17, 31, 57, 105, 193, 355, 653, 1201, 2209, 4063, 7473, 13745, 25281, and 46499.