def disemvowel(text):
To "disemvowel" some piece of text to obfuscate its meaning but still potentially allowing somebody to decipher it provided that they really want to put in the effort to unscramble it, we simply remove every vowel character from that string, and return what remains. For example, disemvoweling the string 'hello world' would produce 'hll wrld'. The letters aeiou are vowels as usual, but to make this problem more interesting, we will use a crude approximation of the correct rule to handle the problematic letter y by defining that the letter y is a vowel whenever the text contains one of the proper vowels aeiou immediately to the left or to the right of that y.