def expand_intervals(intervals):
An interval of consecutive positive integers can be succinctly described as a string that contains its first and last value, inclusive, separated by a minus sign. (This problem is restricted to positive integers so that there can be no ambiguity between the minus sign character used as a separator and an actual unary minus sign in front of an integer.) For example, the interval that contains the numbers 5, 6, 7, 8, 9 could be more concisely described as '5-9'. Multiple intervals can be described together by separating their descriptions with commas. An interval that contains only one value is given as only that value.